requirements.txt
# Django
Django==3.2

# Rest Framework
requests==2.23.0
djangorestframework==3.12.0
django-oauth-toolkit==1.3.2
drf-yasg==1.20.0

# Celery
celery==4.4.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
redis==3.5.2
django-redis==4.11.0
kombu==4.6.8
